requirements.txt
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1
redis>=3.5
celery>=5.0
djangorestframework
slackclient